requirements.txt
internetarchive>=5.5.0
gdown>=5.2.0
google-api-python-client>=2.179.0
google-auth>=2.40.3
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2

# other packges for file date extraction
Pillow>=11.3.0
mutagen>=1.47.0
pypdf>=5.5.0
python-dateutil>=2.9.0.post0
pymediainfo>=7.0.1